+++ CHANGELOG.md
@@ -1,3 +1,33 @@
+## 1.20.0 (2026-07-24)
+
+* feat(list): add button to publish best discs list as WordPress draft (#144) ([8564053](https://github.com/riffvalley/riffvalley-app-front/commit/8564053)), closes [#144](https://github.com/riffvalley/riffvalley-app-front/issues/144)
+* Feat/reunion create as calendar event (#142) ([3e3e295](https://github.com/riffvalley/riffvalley-app-front/commit/3e3e295)), closes [#142](https://github.com/riffvalley/riffvalley-app-front/issues/142)
+
+## 1.19.0 (2026-06-19)
+
+* feat(discs): destacar discos en home y botón pinned en calendario (#141) ([d16d2cb](https://github.com/riffvalley/riffvalley-app-front/commit/d16d2cb)), closes [#141](https://github.com/riffvalley/riffvalley-app-front/issues/141)
+
+## 1.18.0 (2026-06-19)
+
+* feat(reunions): crear reunión vía evento de calendario (#140) ([5743885](https://github.com/riffvalley/riffvalley-app-front/commit/5743885)), closes [#140](https://github.com/riffvalley/riffvalley-app-front/issues/140)
+
+## <small>1.17.2 (2026-06-05)</small>
+
+* Merge branch 'main' of https://github.com/riffvalley/riffvalley-app-front ([1cfb9f6](https://github.com/riffvalley/riffvalley-app-front/commit/1cfb9f6))
+* fix(chat): delete "(editado)" in all comments ([9d17c71](https://github.com/riffvalley/riffvalley-app-front/commit/9d17c71))
+
+## <small>1.17.1 (2026-06-05)</small>
+
+* fix(homepage): delete discs no cover and no votes ([d33a4a7](https://github.com/riffvalley/riffvalley-app-front/commit/d33a4a7))
+
+## 1.17.0 (2026-06-03)
+
+* feat(reunions): crear reunión vía evento de calendario (#139) ([3fac233](https://github.com/riffvalley/riffvalley-app-front/commit/3fac233)), closes [#139](https://github.com/riffvalley/riffvalley-app-front/issues/139)
+* fix background button Todos los discos dark mode ([8a69977](https://github.com/riffvalley/riffvalley-app-front/commit/8a69977))
+* fix login format ([92facfa](https://github.com/riffvalley/riffvalley-app-front/commit/92facfa))
+* fix spoiler votes cover art ([eff7398](https://github.com/riffvalley/riffvalley-app-front/commit/eff7398))
+* fix UI patch notes ([6d0efe4](https://github.com/riffvalley/riffvalley-app-front/commit/6d0efe4))
+
 ## 1.16.0 (2026-05-21)
 
 * add dark mode calendar and calendarbaby + suggest and petitions ([52f2c87](https://github.com/riffvalley/riffvalley-app-front/commit/52f2c87))